Keep maths types in helpers.py for compatibility
